fleshing things out
