--- src/dwn-json-rpc-js/dwn-server-copy-pasta/json-rpc-handlers/dwn/index.ts
@@ -1,1 +1,0 @@
-export * from './process-message.js'
