--- src/dwn-json-rpc-js/dwn-server-copy-pasta/lib/json-rpc-router.ts
@@ -1,34 +1,0 @@
-import type { Dwn } from '@tbd54566975/dwn-sdk-js'
-import type { Readable } from 'node:stream'
-import type { JsonRpcRequest, JsonRpcResponse } from './json-rpc.js'
-
-export type RequestContext = {
-  dwn: Dwn
-  transport: 'http' | 'ws'
-  dataStream?: Readable
-}
-
-export type HandlerResponse = {
-  jsonRpcResponse: JsonRpcResponse,
-  dataStream?: Readable
-}
-
-export type JsonRpcHandler = (JsonRpcRequest: JsonRpcRequest, context: RequestContext) => Promise<HandlerResponse>
-
-export class JsonRpcRouter {
-  private methodHandlers: { [method: string]: JsonRpcHandler }
-
-  constructor() {
-    this.methodHandlers = {}
-  }
-
-  on(methodName: string, handler: JsonRpcHandler) {
-    this.methodHandlers[methodName] = handler
-  }
-
-  async handle(rpcRequest: JsonRpcRequest, context: RequestContext) {
-    const handler = this.methodHandlers[rpcRequest.method]
-
-    return await handler(rpcRequest, context)
-  }
-}
