--- src/dwn-json-rpc-js/index.ts
@@ -1,2 +1,0 @@
-export * from './dwn-json-rpc.js'
-export * from './dwn-server.js'
